+++ annimate_desktop/src/components/main-page/export-trigger.tsx
@@ -6,8 +6,7 @@ import {
   TooltipTrigger,
 } from '@/components/ui/tooltip';
 import { useToast } from '@/components/ui/use-toast';
-import { openPath, revealItemInDir, save } from '@/lib/api';
-import { ExportFormat } from '@/lib/api-types';
+import { documentDir, openPath, revealItemInDir, save } from '@/lib/api';
 import { CancellableOperationError } from '@/lib/mutations';
 import {
   useCanExport,
@@ -17,10 +16,7 @@ import {
 import { formatPercentage } from '@/lib/utils';
 import { File, Folder, Hourglass, X } from 'lucide-react';
 
-const EXPORT_FORMAT_FILTERS: Record<
-  ExportFormat,
-  { name: string; extensions: string[] }[]
-> = {
+const EXPORT_FORMAT_FILTERS = {
   csv: [{ name: 'Comma-separated values (*.csv)', extensions: ['csv'] }],
   xlsx: [{ name: 'Excel (*.xlsx)', extensions: ['xlsx'] }],
 };
@@ -35,7 +31,7 @@ export const ExportTrigger = () => {
     cancelRequested,
     requestCancel,
   } = useExportMatches();
-  const { toast } = useToast();
+  const { dismiss: dismissToast, toast } = useToast();
 
   return (
     <div className="flex-1">
@@ -85,10 +81,14 @@ export const ExportTrigger = () => {
           onClick={async () => {
             const exportFormat = getExportFormat();
             const outputFile = await save({
+              defaultPath: await documentDir(),
               filters: EXPORT_FORMAT_FILTERS[exportFormat],
               title: 'Export to file',
             });
+
             if (outputFile !== null) {
+              dismissToast();
+
               exportMatches(
                 { outputFile },
                 {