+++ annimate_desktop/src/components/main-page/main-page.tsx
@@ -6,32 +6,40 @@ import {
   ResizablePanel,
   ResizablePanelGroup,
 } from '@/components/ui/resizable';
-import { FC } from 'react';
+import { useToast } from '@/components/ui/use-toast';
+import { FC, useEffect } from 'react';
 
-export const MainPage: FC = () => (
-  <ResizablePanelGroup direction="horizontal">
-    <ResizablePanel className="pt-2" defaultSize={30} minSize={30}>
-      <ResizablePanelGroup direction="vertical">
-        <ResizablePanel
-          className="px-3 pt-1 pb-4"
-          defaultSize={50}
-          minSize={30}
-        >
-          <QueryInput />
-        </ResizablePanel>
+export const MainPage: FC = () => {
+  const { dismiss: dismissToast } = useToast();
+  useEffect(() => {
+    dismissToast();
+  }, [dismissToast]);
 
-        <ResizableHandle withHandle />
+  return (
+    <ResizablePanelGroup direction="horizontal">
+      <ResizablePanel className="pt-2" defaultSize={30} minSize={30}>
+        <ResizablePanelGroup direction="vertical">
+          <ResizablePanel
+            className="px-3 pt-1 pb-4"
+            defaultSize={50}
+            minSize={30}
+          >
+            <QueryInput />
+          </ResizablePanel>
 
-        <ResizablePanel className="p-3" defaultSize={50} minSize={30}>
-          <CorpusList />
-        </ResizablePanel>
-      </ResizablePanelGroup>
-    </ResizablePanel>
+          <ResizableHandle withHandle />
 
-    <ResizableHandle withHandle />
+          <ResizablePanel className="p-3" defaultSize={50} minSize={30}>
+            <CorpusList />
+          </ResizablePanel>
+        </ResizablePanelGroup>
+      </ResizablePanel>
 
-    <ResizablePanel className="px-3 pt-2 pb-4" defaultSize={70} minSize={30}>
-      <ExportSection />
-    </ResizablePanel>
-  </ResizablePanelGroup>
-);
+      <ResizableHandle withHandle />
+
+      <ResizablePanel className="px-3 pt-2 pb-4" defaultSize={70} minSize={30}>
+        <ExportSection />
+      </ResizablePanel>
+    </ResizablePanelGroup>
+  );
+};