+++ annimate_desktop/src/components/manage-page/import-trigger.tsx
@@ -9,7 +9,7 @@ import {
   DropdownMenuTrigger,
 } from '@/components/ui/dropdown-menu';
 import { useToast } from '@/components/ui/use-toast';
-import { open } from '@/lib/api';
+import { downloadDir, open } from '@/lib/api';
 import { OpenDialogOptions } from '@/lib/api-types';
 import { useAddCorporaToSet, useImportCorpora } from '@/lib/store';
 import { File, Folder, FolderInput } from 'lucide-react';
@@ -39,7 +39,10 @@ export const ImportTrigger: FC<ImportTriggerProps> = ({
   const [dialogOpen, setDialogOpen, dialogKey] = useDialogState();
 
   const importCorporaFromDialog = async (options: OpenDialogOptions) => {
-    const pathsRaw = await open(options);
+    const pathsRaw = await open({
+      defaultPath: await downloadDir(),
+      ...options,
+    });
 
     if (pathsRaw !== null) {
       importCorpora({ paths: Array.isArray(pathsRaw) ? pathsRaw : [pathsRaw] });